requirements.txt
Django==4.2.20
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
django-allauth==0.61.1
django-environ==0.11.2
Pillow==10.2.0
stripe==7.6.0
supabase==2.0.3
requests==2.31.0
python-telegram-bot==20.5
django-cors-headers==4.3.1
djangorestframework==3.14.0
django-filter==23.5
drf-yasg>=1.21.5
python-dotenv==1.0.1
django-storages==1.14.2
boto3==1.34.34
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
django-redis==5.4.0
celery==5.3.6
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-silk==5.1.0
django-prometheus==2.3.1
prometheus-client==0.19.0
dnspython==2.6.1
httpx>=0.24.0,<0.25.0 